requirements.txt
python-telegram-bot==22.5
openai==1.57.4
aiosqlite==0.21.0
pydantic==2.12.5
pydantic-settings==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
tenacity==9.0.0
httpx==0.28.1
